Implement react router / Project cleanup (#77)
* Implement react-router. Implement basic page design

* Add VoterInfoForm to VoterRegLookup

* Delete modals

* Remove npm commands from contributing guide
