+++ .github/CONTRIBUTING.md
@@ -22,12 +22,12 @@ $ cd GoVote
 
 The project has been properly cloned at this point, and it is time to get it up and running. Run the following commands to build the govote image, create and seed the database, and start the containers.
 ``` bash
-$ npm run docker:build # Builds the govote image.
-$ npm run docker:dev # Starts the services. You can now access the project at http://localhost:3000
-$ npm run docker:createTables # Creates tables
-$ npm run docker:seed # Pulls data from the NC Board of Elections site into the db.
+docker-compose build
+docker-compose up -d # Starts the database and server in the background. Access the app at http://localhost:3000
+docker-compose exec govote node ./bin/create-tables.js # Creates the database tables
+docker-compose exec govote node ./bin/etl.js # Seeds the database
 ```
-To stop the project run `npm run docker:stop`
+To stop the project run `docker-compose stop`
 
 ### Project Details
 The project uses [ReactJS](https://reactjs.org/) as a UI library, and the server runs on NodeJS utilizing the [Express](https://expressjs.com/) framework. Our code linting is done with [ESLint](https://eslint.org/), and if you have followed the Getting Started steps, any linting errors should be present in your console. Please confirm that all linter errors are resolved before [Creating a Pull Request](https://help.github.com/articles/creating-a-pull-request/).
