+++ package.json
@@ -36,7 +36,7 @@
     "react-render-html": "^0.1.6",
     "react-router": "^3.0.2",
     "react-router-bootstrap": "^0.23.1",
-    "react-router-dom": "^4.1.2",
+    "react-router-dom": "^4.3.1",
     "request": "^2.85.0",
     "sw-precache-webpack-plugin": "^0.11.4",
     "unzip": "^0.1.11",
