+++ src/components/App.jsx
@@ -1,141 +1,35 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
+import { BrowserRouter as Router, Link, Route } from 'react-router-dom';
+import { Row, Col, Grid } from 'react-bootstrap';
 
-import { getGISData, getVoterInfo } from '../actions';
-import AboutModal from './AboutModal';
 import Header from './Header';
-import MapContainer from './MapContainer';
-import RegistrationInfoModal from './RegistrationInfoModal';
-import VerifyVoterInfo from './VerifyVoterInfo';
-import VoterInfoForm from './VoterInfoForm';
-import VoterModal from './VoterModal';
-
+import Home from './pages/Home';
+import About from './pages/About';
+import VoterRegLookup from './pages/VoterRegLookup';
 
 class App extends Component {
-
-  state = {
-    voterModalShow: true,
-    regInfoModalShow: false,
-    aboutModalShow: false,
-    firstName: '',
-    lastName: '',
-    voterInfo: [],
-    isloading: false,
-    voterAddress: '',
-    layers: {
-      commissionerDist: null,
-    },
-    firstNameValidationState: null,
-    lastNameValidationState: null,
-    formErrors: [],
-    voterInfoFound: true,
-  };
-
-  componentDidMount() {
-    getGISData()
-      .then(layers => this.setState({ layers }));
-  }
-
-  _getVoterInfo = () => {
-    if (this._validateVoterInput()) {
-      this.setState({
-        firstNameValidationState: null,
-        lastNameValidationState: null,
-      });
-
-      this.setState({ isloading: true });
-
-      getVoterInfo(this.state.firstName, this.state.lastName)
-        .then((voterInfo) => {
-          this.setState({
-            voterInfo,
-            voterInfoFound: voterInfo.length > 0,
-            isloading: false,
-          });
-        });
-    } else {
-      this.setState({ voterInfo: [] });
-    }
-  }
-
-  _validateVoterInput = () => {
-    let result = true;
-    const errors = [];
-    if (this.state.firstName === '') {
-      this.setState({ firstNameValidationState: 'error' });
-      errors.push('first name');
-      result = false;
-    }
-    if (this.state.lastName === '') {
-      this.setState({ lastNameValidationState: 'error' });
-      errors.push('last name');
-      result = false;
-    }
-    this.setState({ formErrors: errors });
-    return result;
-  }
-
-  _getVoterAddress = (address) => {
-    this.setState({ voterAddress: address });
-  }
-
-  _showVoterInfoModal = () => {
-    this.setState({
-      regInfoModalShow: false,
-      voterModalShow: true,
-    });
-  }
-
-  _showRegInfoModal = () => {
-    this.setState({
-      regInfoModalShow: true,
-      voterModalShow: false,
-      voterInfo: [],
-    });
-  }
-
-  _handleInputChange = (event) => {
-    const { name, value } = event.target;
-
-    this.setState({
-      [name]: value,
-      voterInfo: [],
-      selectedAddress: {},
-      showUserStatus: false,
-    });
-  }
-
-  _handleShowAboutModal = () => {
-    this.setState({ aboutModalShow: true });
-  }
-
-  _handleHideAboutModal = () => {
-    this.setState({ aboutModalShow: false });
-  }
-
   render() {
-    const voterModalShow = () => this.setState({
-      voterModalShow: false,
-      voterInfo: [],
-      formErrors: [],
-      firstNameValidationState: null,
-      lastNameValidationState: null,
-    });
-    const regInfoModalShow = () => this.setState({ regInfoModalShow: false });
-
     return (
-      this.state.layers.length > 0 &&
+      <Router>
         <div className="app__wrapper">
-          <Header showVoterInfoModal={this._showVoterInfoModal} showAboutModal={this._handleShowAboutModal} />
-          <MapContainer layers={this.state.layers} />
-
-          <AboutModal show={this.state.aboutModalShow} hide={this._handleHideAboutModal} onClick={this._handleShowAboutModal} />
-          <RegistrationInfoModal show={this.state.regInfoModalShow} hide={regInfoModalShow} />
-          <VoterModal show={this.state.voterModalShow} hide={voterModalShow}>
-            <VoterInfoForm onSubmit={this._getVoterInfo} onUpdate={this._handleInputChange} voterInfoFound={this.state.voterInfoFound} firstNameValidationState={this.state.firstNameValidationState} lastNameValidationState={this.state.lastNameValidationState} formErrors={this.state.formErrors} isLoading={this.state.isloading} />
-            <VerifyVoterInfo voterInfo={this.state.voterInfo} showRegInfoModal={this._showRegInfoModal} voterAddress={this._getVoterAddress} />
-          </VoterModal>
+          <Header>
+            <Link to="/" className="btn btn-link header__details-action">Home</Link>
+            <Link to="/about" className="btn btn-link header__details-action">About</Link>
+          </Header>
+          <div className="app__body">
+            <Grid >
+              <Row>
+                <Col md={8} mdOffset={2}>
+                  <Route exact path="/" component={Home} />
+                  <Route path="/about" component={About} />
+                  <Route path="/voter-lookup" component={VoterRegLookup} />
+                </Col>
+              </Row>
+            </Grid>
+          </div>
         </div>
+      </Router>
     );
   }
 }
