+++ src/style.css
@@ -7,12 +7,18 @@ html, body, #app {
   padding: 0;
   height: 100%;
   width: 100%;
-  background-color: #283944;
+  background-color: white;
 }
 
 .app__wrapper {
   display: flex;
   flex-direction: column;
+  height: 100%;
+}
+
+.app__body {
+  background: white;
+  width: 100%;
   height: 100%;
 }
 
@@ -25,6 +31,7 @@ html, body, #app {
   display: flex;
   align-items: center;
   justify-content: space-between;
+  background-color: #283944;
 }
 
 .header__title {
